requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
Django==4.1.3
django-cors-headers==3.13.0
django-jazzmin==2.6.0
django-redis==5.2.0
djangorestframework==3.14.0
idna==3.4
packaging==21.3
Pillow==9.3.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.6
qrcode==7.3.1
razorpay==1.3.0
redis==4.3.4
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
wrapt==1.14.1
